Create utils/i18n.py
