+++ wagtail/contrib/frontend_cache/utils.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.module_loading import import_string
 
-from wagtail.utils.coreutils import get_content_languages
+import wagtail.utils.i18n
 
 logger = logging.getLogger("wagtail.frontendcache")
 
@@ -68,7 +68,9 @@ def purge_urls_from_cache(urls, backend_settings=None, backends=None):
     # This depends on settings.USE_I18N
     # If WAGTAIL_I18N_ENABLED is True, this defaults to WAGTAIL_CONTENT_LANGUAGES
     wagtail_i18n_enabled = getattr(settings, "WAGTAIL_I18N_ENABLED", False)
-    content_languages = get_content_languages() if wagtail_i18n_enabled else {}
+    content_languages = (
+        wagtail.utils.i18n.get_content_languages() if wagtail_i18n_enabled else {}
+    )
     languages = getattr(
         settings, "WAGTAILFRONTENDCACHE_LANGUAGES", list(content_languages.keys())
     )
